# wikify-web/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wikify Web API",
    "description": "AI-powered repository documentation and chat system",
    "contact": {
      "name": "Wikify Team",
      "email": "support@wikify.dev"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://api.wikify.dev",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Refresh access token",
        "description": "Refresh access token using a valid refresh token",
        "operationId": "refresh_token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "401": {
            "description": "Invalid or expired refresh token"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/chat": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Ask a question",
        "description": "Ask a question about the repository using RAG (Retrieval-Augmented Generation)",
        "operationId": "chat_query",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Question answered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatQueryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/stream": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Stream chat response",
        "description": "Ask a question and receive streaming response (not yet implemented)",
        "operationId": "chat_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Streaming response started"
          },
          "501": {
            "description": "Not implemented"
          }
        }
      }
    },
...
    "/api/history/{repository_id}": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Get query history",
        "description": "Get chat history for a specific repository (requires SQLite feature)",
        "operationId": "get_query_history",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "description": "Repository ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Query history retrieved successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TokenPair",
            "description": "Authentication tokens"
          },
          {
            "type": "object",
            "required": [
              "user"
            ],
            "properties": {
              "user": {
                "$ref": "#/components/schemas/UserInfo",
                "description": "User information"
              }
            }
          }
        ],
        "description": "User registration/login response"
      },
...
      "ChatQueryRequest": {
        "type": "object",
        "description": "Chat query request",
        "required": [
          "repository_id",
          "question"
        ],
        "properties": {
          "context": {
            "type": [
              "string",
              "null"
            ]
          },
          "max_results": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Maximum number of results to return",
            "minimum": 0
          },
          "question": {
            "type": "string",
            "example": "How does the authentication work?"
          },
          "repository_id": {
            "type": "string",
            "example": "repo-uuid-string"
          }
        }
      },
      "ChatQueryResponse": {
        "type": "object",
        "description": "Chat query response",
        "required": [
          "answer",
          "sources",
          "repository_id",
          "timestamp"
        ],
        "properties": {
          "answer": {
            "type": "string"
          },
          "repository_id": {
            "type": "string",
            "example": "repo-uuid-string"
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceDocument"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "RefreshRequest": {
        "type": "object",
        "description": "Token refresh request",
        "required": [
          "refresh_token"
        ],
        "properties": {
          "refresh_token": {
            "type": "string",
            "description": "Refresh token to exchange for new access token",
            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
          }
        }
      },
...
      "SourceDocument": {
        "type": "object",
        "description": "Source document information",
        "required": [
          "file_path",
          "content",
          "similarity_score"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "file_path": {
            "type": "string",
            "example": "src/main.rs"
          },
          "similarity_score": {
            "type": "number",
            "format": "double",
            "example": 0.85
          }
        }
      },
...
      "TokenPair": {
        "type": "object",
        "description": "JWT token pair (access + refresh)",
        "required": [
          "access_token",
          "refresh_token",
          "token_type",
          "expires_in"
        ],
        "properties": {
          "access_token": {
            "type": "string",
            "description": "Access token for API requests",
            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
          },
          "expires_in": {
            "type": "integer",
            "format": "int64",
            "description": "Token expiration time in seconds",
            "example": 3600
          },
          "refresh_token": {
            "type": "string",
            "description": "Refresh token for obtaining new access tokens",
            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
          },
          "token_type": {
            "type": "string",
            "description": "Token type (always \"Bearer\")",
            "example": "Bearer"
          }
        }
      },
      "UserInfo": {
        "type": "object",
        "description": "Public user information",
        "required": [
          "id",
          "username",
          "email",
          "permissions",
          "is_admin",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Account creation timestamp",
            "example": "2024-01-01T00:00:00Z"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Display name",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "john@example.com"
          },
          "id": {
            "type": "string",
            "description": "Unique user identifier",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "is_admin": {
            "type": "boolean",
            "description": "Whether user is admin",
            "example": false
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "User permissions",
            "example": [
              "GenerateWiki",
              "DeepResearch"
            ]
          },
          "username": {
            "type": "string",
            "description": "Username",
            "example": "john_doe"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}